+++ src/solc/mod.rs
@@ -8,8 +8,11 @@ pub mod standard_json;
 pub mod version;
 
 use std::collections::BTreeMap;
+use std::collections::HashMap;
 use std::io::Write;
 use std::path::PathBuf;
+use std::sync::OnceLock;
+use std::sync::RwLock;
 
 use self::combined_json::CombinedJson;
 use self::pipeline::Pipeline;
@@ -21,11 +24,12 @@ use self::version::Version;
 ///
 /// The Solidity compiler.
 ///
+#[derive(Debug, Clone)]
 pub struct Compiler {
     /// The binary executable name.
     pub executable: String,
-    /// The lazily-initialized compiler version.
-    pub version: Option<Version>,
+    /// The `solc` compiler version.
+    pub version: Version,
 }
 
 impl Compiler {
@@ -48,37 +52,49 @@ impl Compiler {
     pub const LAST_SUPPORTED_VERSION: semver::Version = semver::Version::new(0, 8, 25);
 
     ///
-    /// A shortcut constructor.
+    /// A shortcut constructor lazily using a thread-safe cell.
     ///
     /// Different tools may use different `executable` names. For example, the integration tester
     /// uses `solc-<version>` format.
     ///
-    pub fn new(executable: String) -> anyhow::Result<Self> {
-        if let Err(error) = which::which(executable.as_str()) {
+    pub fn new(executable: &str) -> anyhow::Result<Self> {
+        if let Some(executable) = Self::executables()
+            .read()
+            .expect("Sync")
+            .get(executable)
+            .cloned()
+        {
+            return Ok(executable);
+        }
+        let mut executables = Self::executables().write().expect("Sync");
+
+        if let Err(error) = which::which(executable) {
             anyhow::bail!(
                 "The `{executable}` executable not found in ${{PATH}}: {}",
                 error
             );
         }
-        Ok(Self {
-            executable,
-            version: None,
-        })
+        let version = Self::parse_version(executable)?;
+        let compiler = Self {
+            executable: executable.to_owned(),
+            version,
+        };
+
+        executables.insert(executable.to_owned(), compiler.clone());
+        Ok(compiler)
     }
 
     ///
     /// The Solidity `--standard-json` mirror.
     ///
     pub fn standard_json(
-        &mut self,
+        &self,
         mut input: StandardJsonInput,
         pipeline: Option<Pipeline>,
         base_path: Option<String>,
         include_paths: Vec<String>,
         allow_paths: Option<String>,
     ) -> anyhow::Result<StandardJsonOutput> {
-        let version = self.version()?;
-
         let mut command = std::process::Command::new(self.executable.as_str());
         command.stdin(std::process::Stdio::piped());
         command.stdout(std::process::Stdio::piped());
@@ -140,7 +156,7 @@ impl Compiler {
 
         if let Some(pipeline) = pipeline {
             let suppressed_warnings = input.suppressed_warnings.take().unwrap_or_default();
-            output.preprocess_ast(&version, pipeline, suppressed_warnings.as_slice())?;
+            output.preprocess_ast(&self.version, pipeline, suppressed_warnings.as_slice())?;
         }
         output.remove_evm();
 
@@ -233,11 +249,11 @@ impl Compiler {
     /// Validates the Yul project as paths and libraries.
     ///
     pub fn validate_yul_paths(
-        &mut self,
+        &self,
         paths: &[PathBuf],
         libraries: BTreeMap<String, BTreeMap<String, String>>,
     ) -> anyhow::Result<StandardJsonOutput> {
-        if self.version()?.default != Self::LAST_SUPPORTED_VERSION {
+        if self.version.default != Self::LAST_SUPPORTED_VERSION {
             anyhow::bail!(
                 "Yul validation is only supported with the latest supported version of the Solidity compiler: {}",
                 Self::LAST_SUPPORTED_VERSION,
@@ -256,10 +272,10 @@ impl Compiler {
     /// Validates the Yul project as standard JSON input.
     ///
     pub fn validate_yul_standard_json(
-        &mut self,
+        &self,
         mut solc_input: StandardJsonInput,
     ) -> anyhow::Result<StandardJsonOutput> {
-        if self.version()?.default != Self::LAST_SUPPORTED_VERSION {
+        if self.version.default != Self::LAST_SUPPORTED_VERSION {
             anyhow::bail!(
                 "Yul validation is only supported with the latest supported version of the Solidity compiler: {}",
                 Self::LAST_SUPPORTED_VERSION,
@@ -285,22 +301,26 @@ impl Compiler {
     }
 
     ///
+    /// Returns the global shared array of `solc` executables.
+    ///
+    fn executables() -> &'static RwLock<HashMap<String, Self>> {
+        static EXECUTABLES: OnceLock<RwLock<HashMap<String, Compiler>>> = OnceLock::new();
+        EXECUTABLES.get_or_init(|| RwLock::new(HashMap::new()))
+    }
+
+    ///
     /// The `solc --version` mini-parser.
     ///
-    pub fn version(&mut self) -> anyhow::Result<Version> {
-        if let Some(version) = self.version.as_ref() {
-            return Ok(version.to_owned());
-        }
-
-        let mut command = std::process::Command::new(self.executable.as_str());
+    fn parse_version(executable: &str) -> anyhow::Result<Version> {
+        let mut command = std::process::Command::new(executable);
         command.arg("--version");
-        let output = command.output().map_err(|error| {
-            anyhow::anyhow!("{} subprocess error: {:?}", self.executable, error)
-        })?;
+        let output = command
+            .output()
+            .map_err(|error| anyhow::anyhow!("{} subprocess error: {:?}", executable, error))?;
         if !output.status.success() {
             anyhow::bail!(
                 "{} error: {}",
-                self.executable,
+                executable,
                 String::from_utf8_lossy(output.stderr.as_slice()).to_string()
             );
         }
@@ -309,26 +329,22 @@ impl Compiler {
         let long = stdout
             .lines()
             .nth(1)
-            .ok_or_else(|| {
-                anyhow::anyhow!("{} version parsing: not enough lines", self.executable)
-            })?
+            .ok_or_else(|| anyhow::anyhow!("{} version parsing: not enough lines", executable))?
             .split(' ')
             .nth(1)
             .ok_or_else(|| {
                 anyhow::anyhow!(
                     "{} version parsing: not enough words in the 2nd line",
-                    self.executable
+                    executable
                 )
             })?
             .to_owned();
         let default: semver::Version = long
             .split('+')
             .next()
-            .ok_or_else(|| {
-                anyhow::anyhow!("{} version parsing: metadata dropping", self.executable)
-            })?
+            .ok_or_else(|| anyhow::anyhow!("{} version parsing: metadata dropping", executable))?
             .parse()
-            .map_err(|error| anyhow::anyhow!("{} version parsing: {}", self.executable, error))?;
+            .map_err(|error| anyhow::anyhow!("{} version parsing: {}", executable, error))?;
 
         let l2_revision: Option<semver::Version> = stdout
             .lines()
@@ -352,8 +368,6 @@ impl Compiler {
                 version.default
             );
         }
-
-        self.version = Some(version.clone());
 
         Ok(version)
     }