+++ src/tests/standard_json.rs
@@ -40,10 +40,10 @@ fn standard_json_yul_default_validated() {
         PathBuf::from("src/tests/examples/standard_json_input/yul_default.json").as_path(),
     ))
     .expect("Standard JSON reading error");
-    let solc_compiler = SolcCompiler::new(SolcCompiler::DEFAULT_EXECUTABLE_NAME.to_owned())
+    let solc_compiler = SolcCompiler::new(SolcCompiler::DEFAULT_EXECUTABLE_NAME)
         .expect("`solc` initialization error");
     let solc_output =
-        super::build_yul_standard_json(solc_input, Some(solc_compiler)).expect("Test failure");
+        super::build_yul_standard_json(solc_input, Some(&solc_compiler)).expect("Test failure");
 
     assert!(!solc_output
         .contracts
@@ -93,10 +93,10 @@ fn standard_json_yul_default_urls_validated() {
         PathBuf::from("src/tests/examples/standard_json_input/yul_default_urls.json").as_path(),
     ))
     .expect("Standard JSON reading error");
-    let solc_compiler = SolcCompiler::new(SolcCompiler::DEFAULT_EXECUTABLE_NAME.to_owned())
+    let solc_compiler = SolcCompiler::new(SolcCompiler::DEFAULT_EXECUTABLE_NAME)
         .expect("`solc` initialization error");
     let solc_output =
-        super::build_yul_standard_json(solc_input, Some(solc_compiler)).expect("Test failure");
+        super::build_yul_standard_json(solc_input, Some(&solc_compiler)).expect("Test failure");
 
     assert!(!solc_output
         .contracts
@@ -146,10 +146,10 @@ fn standard_json_yul_eravm_validated() {
         PathBuf::from("src/tests/examples/standard_json_input/yul_eravm.json").as_path(),
     ))
     .expect("Standard JSON reading error");
-    let solc_compiler = SolcCompiler::new(SolcCompiler::DEFAULT_EXECUTABLE_NAME.to_owned())
+    let solc_compiler = SolcCompiler::new(SolcCompiler::DEFAULT_EXECUTABLE_NAME)
         .expect("`solc` initialization error");
     let solc_output =
-        super::build_yul_standard_json(solc_input, Some(solc_compiler)).expect("Test failure");
+        super::build_yul_standard_json(solc_input, Some(&solc_compiler)).expect("Test failure");
 
     assert!(!solc_output
         .contracts
@@ -199,10 +199,10 @@ fn standard_json_yul_eravm_urls_validated() {
         PathBuf::from("src/tests/examples/standard_json_input/yul_eravm_urls.json").as_path(),
     ))
     .expect("Standard JSON reading error");
-    let solc_compiler = SolcCompiler::new(SolcCompiler::DEFAULT_EXECUTABLE_NAME.to_owned())
+    let solc_compiler = SolcCompiler::new(SolcCompiler::DEFAULT_EXECUTABLE_NAME)
         .expect("`solc` initialization error");
     let solc_output =
-        super::build_yul_standard_json(solc_input, Some(solc_compiler)).expect("Test failure");
+        super::build_yul_standard_json(solc_input, Some(&solc_compiler)).expect("Test failure");
 
     assert!(!solc_output
         .contracts